+++ Makefile
@@ -7,8 +7,9 @@ KERN_OBJ = kernel/main.o \
            kernel/console.o \
            kernel/intr.o \
            kernel/queue.o \
-           kernel/keyboard.o \
-           kernel/util.o
+           kernel/keyboard.o
+
+LIB_OBJ  = lib/string.o
 
 MM_OBJ   = mm/memory.o \
            mm/pgtable.o
@@ -48,7 +49,7 @@ $(ARCH)/head.bin: $(ARCH)/head.o
 	ld $^ -T $(ARCH)/head.ld -Map head.map -o $@
 
 # TODO : ひとまずkernelにすべてリンクするが、あとでユーザ空間を分ける
-kernel/kernel.bin: $(KERN_OBJ) $(BIN_OBJ) $(MM_OBJ)
+kernel/kernel.bin: $(KERN_OBJ) $(LIB_OBJ) $(BIN_OBJ) $(MM_OBJ)
 	ld $^ -T kernel/kernel.ld -Map kernel.map -o $@
 
 run: img
