+++ bin/ls.c
@@ -2,7 +2,6 @@
 
 #include <console.h>
 #include <exit.h>
-#include <util.h>
 
 int ls() {
   printf("hello, %c%s", 'w', "orld!");
